converted fund-service to multiplatform module

+++ service/fund/fund-service/src/main/kotlin/ro/jf/funds/fund/service/mapper/TransactionMapper.kt
@@ -52,7 +52,8 @@ fun TransactionRecord.toTO() = TransactionRecordTO(
     id = id,
     accountId = accountId,
     fundId = fundId,
-    amount = amount,
+    // TODO(Johann) not ok
+    amount = amount.toDouble(),
     unit = unit,
     labels = labels,
 )
